Deprecate `Endpoint.Feed`
